fix bfloat16 error message
Update esm3.py

+++ esm/layers/geom_attention.py
@@ -148,6 +148,5 @@ class GeometricReasoningOriginalImpl(nn.Module):
         )
         if self.mask_and_zero_frameless:
             attn_out = attn_out.masked_fill(~affine_mask[..., None], 0.0)
-        s = self.out_proj(attn_out)
-
+        s = self.out_proj(attn_out.to(torch.bfloat16))
         return s
